Consolidate field validations (#4292)
* First pass at consolidating general validations for HCA only

* remove commented code

* Clean up remaining HCA validations

* Fix name validation

* Fix failing test

* Fix a failing test

* Add requirement for isValidSection prop back
