+++ src/js/hca/components/ReviewCollapsiblePanel.jsx
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import _ from 'lodash';
 
 import { ensureFieldsInitialized, updateIncompleteStatus, updateVerifiedStatus, updateCompletedStatus } from '../actions';
-import * as validations from '../../common/utils/validations';
+import { isValidSection } from '../utils/validations';
 
 const Element = Scroll.Element;
 const scroller = Scroll.scroller;
@@ -42,7 +42,7 @@ class ReviewCollapsiblePanel extends React.Component {
     const sectionFields = this.props.uiData.sections[currentPath].fields;
 
     this.props.onFieldsInitialized(sectionFields);
-    if (validations.isValidSection(currentPath, formData)) {
+    if (isValidSection(currentPath, formData)) {
       this.props.onUpdateSaveStatus(currentPath);
     }
     this.scrollToTop();
